# Conflicts:
#	src/main/java/com/example/demo/domain/service/post/PostService.java
#	src/main/java/com/example/demo/infra/repository/PostRepository.java
#	src/main/java/com/example/demo/infra/repository/UserRepository.java
#	src/main/java/com/example/demo/presentation/controller/pageproperty/SessionKeyword.java
#	src/main/java/com/example/demo/presentation/controller/pageproperty/TransitionTargetPageNameKeyword.java
#	src/main/java/com/example/demo/presentation/form/LoginUserForm.java
#	src/main/java/com/example/demo/presentation/form/customer/CustomerAccountEditForm.java

+++ src/main/java/com/example/demo/presentation/form/post/PostListForm.java
@@ -9,9 +9,9 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class PostListForm {
-
-    
+	/** 投稿Id　*/
 	private Integer postId;
+
 	/** 投稿者種別（artisan / customer） */
     private String userType;
 
@@ -31,5 +31,5 @@ public class PostListForm {
     private List<String> tags;
 
     /** いいね数 */
-    private Integer favoriteCount;
+    private long favoriteCount;
 }
